requirements.txt
streamlit==1.36.0
requests>=2.31
python-dotenv>=1.0
pandas>=2.0
numpy>=1.26
